Refactored into sub-module to reduce line length
