+++ benchmarks/analyze_field_regressions/__init__.py
@@ -1,0 +1,40 @@
+from ._aggregate import (
+    _aggregate_model_results,
+    _is_meaningful_label_change,
+    _label_is_relevant,
+    _pair_is_relevant,
+)
+from ._cases import _normalize_for_comparison
+from ._cli import main
+from ._fetch import GROBID_DEFAULT_URL, PARSER_DEFAULT_URL
+from ._loop import DEFAULT_CONCURRENCY, _resolve_concurrency
+from ._models import FIELD_MODEL, MODEL_RELEVANT_LABELS, _get_model_chain
+from ._report import (
+    _feature_section_label,
+    _render_feature_summary_table,
+    _render_feature_table,
+)
+from ._types import FeatureSummary, FieldPresenceSummary, ModelSummary, RegressionCase
+
+__all__ = [
+    'DEFAULT_CONCURRENCY',
+    'FIELD_MODEL',
+    'GROBID_DEFAULT_URL',
+    'MODEL_RELEVANT_LABELS',
+    'PARSER_DEFAULT_URL',
+    'FeatureSummary',
+    'FieldPresenceSummary',
+    'ModelSummary',
+    'RegressionCase',
+    '_aggregate_model_results',
+    '_feature_section_label',
+    '_get_model_chain',
+    '_is_meaningful_label_change',
+    '_label_is_relevant',
+    '_normalize_for_comparison',
+    '_pair_is_relevant',
+    '_render_feature_summary_table',
+    '_render_feature_table',
+    '_resolve_concurrency',
+    'main',
+]
