+++ benchmarks/analyze_field_regressions/__main__.py
@@ -1,0 +1,3 @@
+from ._cli import main
+
+main()
